requirements.txt
# Web App Requirements
flask==3.0.3
joblib==1.4.2
pandas==2.2.2
gunicorn==20.1.0
requests==2.32.3
python-dotenv==0.21.0
streamlit==1.37.1

# Notebook Requirements (commented out to reduce slug size)
scikit-learn==1.5.1
scikit-surprise==1.1.4
# nltk==3.9.1
# numpy==1.26.4
# matplotlib==3.8.4
# seaborn==0.13.2